Add batch metadata operations for Python bindings (#261)
* Add batch metadata operations for Python bindings

- Implement get_multiple_joker_metadata() to get metadata for a list of joker IDs
- Implement get_all_joker_metadata() to get metadata for all registered jokers
- Add Python getters for JokerMetadata fields to make them accessible
- Add comprehensive tests for batch operations
- Temporarily register test jokers in registry for testing

Part of PR split from #221 (issue #172)



* fix: Apply rustfmt formatting fixes

Applied automated formatting to resolve CI rustfmt failures.


* fix: Resolve f64 to usize type conversion after main merge

Fixed compilation error in vouchers module where game.money (f64) was being
assigned to GameState.money (usize) field. Added explicit type conversion.

This resolves the compilation failure introduced by the f64 migration in main.


---------

+++ core/src/vouchers/mod.rs
@@ -437,7 +437,7 @@ impl From<&crate::game::Game> for GameState {
         };
 
         Self {
-            money: game.money,
+            money: game.money as usize,
             ante: ante_value,
             hand_size: 8, // Base hand size, vouchers would modify this
             joker_slots: game.config.joker_slots,
